Making it compile, by removing obsolete code
