+++ samples/FluentMigrator.Example.Migrator/Program.cs
@@ -40,10 +40,6 @@ namespace FluentMigrator.Example.Migrator
             var app = new CommandLineApplication();
 
             var help = app.HelpOption();
-            var mode = app.Option<MigrationMode>(
-                "-m|--mode <MODE>",
-                "The mode of the application (legacy or di)",
-                CommandOptionType.SingleValue);
             var processor = app.Option(
                 "-d|--dialect <DIALECT>",
                 $"The database dialect ({string.Join(",", DefaultConfigurations.Keys)})",
@@ -56,20 +52,8 @@ namespace FluentMigrator.Example.Migrator
             app.OnExecute(
                 () =>
                 {
-                    var selectedMode = mode.HasValue() ? mode.ParsedValue : MigrationMode.DI;
                     var dbConfig = CreateDatabaseConfiguration(processor, connectionString);
-                    switch (selectedMode)
-                    {
-                        case MigrationMode.Legacy:
-                            Console.WriteLine(@"Using legacy mode");
-                            RunInLegacyMode(dbConfig);
-                            break;
-                        case MigrationMode.DI:
-                            Console.WriteLine(@"Using dependency injection");
-                            RunWithServices(dbConfig);
-                            break;
-                    }
-
+                    RunWithServices(dbConfig);
                     return 0;
                 });
 
@@ -108,12 +92,6 @@ namespace FluentMigrator.Example.Migrator
             }
 
             return DefaultDatabaseConfigurations.Sqlite;
-        }
-
-        private enum MigrationMode
-        {
-            Legacy,
-            DI,
         }
     }
 }
