+++ src/FluentMigrator.Runner.Core/FluentMigratorServiceCollectionExtensions.cs
@@ -199,96 +199,6 @@ namespace Microsoft.Extensions.DependencyInjection
             return services;
         }
 
-        /// <summary>
-        /// Creates services for a given runner context, connection string provider and assembly loader factory.
-        /// </summary>
-        /// <param name="runnerContext">The runner context</param>
-        /// <param name="connectionStringProvider">The connection string provider</param>
-        /// <param name="defaultAssemblyLoaderFactory">The assembly loader factory</param>
-        /// <returns>The new service collection</returns>
-        [NotNull]
-        [Obsolete]
-        internal static IServiceCollection CreateServices(
-            [NotNull] this IRunnerContext runnerContext,
-            [CanBeNull] IConnectionStringProvider connectionStringProvider,
-            [CanBeNull] AssemblyLoaderFactory defaultAssemblyLoaderFactory = null)
-        {
-            var services = new ServiceCollection();
-            var assemblyLoaderFactory = defaultAssemblyLoaderFactory ?? new AssemblyLoaderFactory();
-
-            if (!runnerContext.NoConnection && connectionStringProvider == null)
-            {
-                runnerContext.NoConnection = true;
-            }
-
-            // Configure the migration runner
-            services
-                .AddLogging(lb => lb.AddProvider(new LegacyFluentMigratorLoggerProvider(runnerContext.Announcer)))
-                .AddFluentMigratorCore()
-                .AddAllDatabases()
-                .Configure<SelectingProcessorAccessorOptions>(opt => opt.ProcessorId = runnerContext.Database)
-                .AddSingleton(assemblyLoaderFactory)
-                .Configure<TypeFilterOptions>(
-                    opt =>
-                    {
-                        opt.Namespace = runnerContext.Namespace;
-                        opt.NestedNamespaces = runnerContext.NestedNamespaces;
-                    })
-                .Configure<AssemblySourceOptions>(opt => opt.AssemblyNames = runnerContext.Targets)
-                .Configure<RunnerOptions>(
-                    opt => { opt.SetValuesFrom(runnerContext); })
-                .Configure<ProcessorOptions>(opt => { opt.SetValuesFrom(runnerContext); })
-                .Configure<AppConfigConnectionStringAccessorOptions>(
-                    opt => opt.ConnectionStringConfigPath = runnerContext.ConnectionStringConfigPath);
-
-            // Configure the processor
-            if (runnerContext.NoConnection)
-            {
-                // Always return the connectionless processor
-                services
-                    .AddScoped<IProcessorAccessor, ConnectionlessProcessorAccessor>();
-            }
-
-            return services;
-        }
-
-        /// <summary>
-        /// Add all database services to the service collection
-        /// </summary>
-        /// <param name="services">The service collection</param>
-        /// <returns>The service collection</returns>
-        private static IServiceCollection AddAllDatabases(this IServiceCollection services)
-        {
-            return services
-                .ConfigureRunner(
-                    builder => builder
-                        .AddDb2()
-                        .AddDb2ISeries()
-                        .AddDotConnectOracle()
-                        .AddDotConnectOracle12C()
-                        .AddFirebird()
-                        .AddHana()
-                        .AddMySql4()
-                        .AddMySql5()
-                        .AddOracle()
-                        .AddOracle12C()
-                        .AddOracleManaged()
-                        .AddOracle12CManaged()
-                        .AddPostgres()
-                        .AddPostgres92()
-                        .AddRedshift()
-                        .AddSqlAnywhere()
-                        .AddSQLite()
-                        .AddSqlServer()
-                        .AddSqlServer2000()
-                        .AddSqlServer2005()
-                        .AddSqlServer2008()
-                        .AddSqlServer2012()
-                        .AddSqlServer2014()
-                        .AddSqlServer2016()
-                        .AddSqlServerCe());
-        }
-
         private class MigrationRunnerBuilder : IMigrationRunnerBuilder
         {
             public MigrationRunnerBuilder(IServiceCollection services)
