+++ src/FluentMigrator.Runner.Core/MigrationRunner.cs
@@ -80,106 +80,6 @@ namespace FluentMigrator.Runner
         /// <summary>
         /// Initializes a new instance of the <see cref="MigrationRunner"/> class.
         /// </summary>
-        /// <param name="assembly">The assembly to scan for migrations, etc...</param>
-        /// <param name="runnerContext">The runner context</param>
-        /// <param name="processor">The migration processor</param>
-        [Obsolete]
-        public MigrationRunner([NotNull] Assembly assembly, [NotNull] IRunnerContext runnerContext, [NotNull] IMigrationProcessor processor)
-            : this(assembly, runnerContext, processor, conventionSet: null)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="MigrationRunner"/> class.
-        /// </summary>
-        /// <param name="assembly">The assembly to scan for migrations, etc...</param>
-        /// <param name="runnerContext">The runner context</param>
-        /// <param name="processor">The migration processor</param>
-        /// <param name="conventionSet">The expression convention set</param>
-        [Obsolete]
-        public MigrationRunner([NotNull] Assembly assembly, [NotNull] IRunnerContext runnerContext, [NotNull] IMigrationProcessor processor, [CanBeNull] IConventionSet conventionSet)
-            : this(new SingleAssembly(assembly), runnerContext, processor, versionTableMetaData: null, migrationRunnerConventions: null, conventionSet, migrationScopeHandler: null)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="MigrationRunner"/> class.
-        /// </summary>
-        /// <param name="assemblies">The collection of assemblies to scan for migrations, etc...</param>
-        /// <param name="runnerContext">The runner context</param>
-        /// <param name="processor">The migration processor</param>
-        /// <param name="versionTableMetaData">The version table metadata</param>
-        /// <param name="migrationRunnerConventions">The custom migration runner conventions</param>
-        [Obsolete]
-        public MigrationRunner(
-            [NotNull] IAssemblyCollection assemblies, [NotNull] IRunnerContext runnerContext,
-            [NotNull] IMigrationProcessor processor, IVersionTableMetaData versionTableMetaData = null,
-            [CanBeNull] IMigrationRunnerConventions migrationRunnerConventions = null)
-            : this(assemblies, runnerContext, processor, versionTableMetaData, migrationRunnerConventions, conventionSet: null, migrationScopeHandler: null)
-        {
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="MigrationRunner"/> class.
-        /// </summary>
-        /// <param name="assemblies">The collection of assemblies to scan for migrations, etc...</param>
-        /// <param name="runnerContext">The runner context</param>
-        /// <param name="processor">The migration processor</param>
-        /// <param name="versionTableMetaData">The version table metadata</param>
-        /// <param name="migrationRunnerConventions">The custom migration runner conventions</param>
-        /// <param name="conventionSet">The expression convention set</param>
-        /// <param name="migrationScopeHandler">The migration scope handler</param>
-        [Obsolete]
-        public MigrationRunner(
-            [NotNull] IAssemblyCollection assemblies, [NotNull] IRunnerContext runnerContext,
-            [NotNull] IMigrationProcessor processor, [CanBeNull] IVersionTableMetaData versionTableMetaData,
-            [CanBeNull] IMigrationRunnerConventions migrationRunnerConventions, [CanBeNull] IConventionSet conventionSet,
-            [CanBeNull] IMigrationScopeManager migrationScopeHandler = null)
-        {
-            _migrationAssemblies = assemblies;
-            _logger = new AnnouncerFluentMigratorLogger(runnerContext.Announcer);
-            _stopWatch = runnerContext.StopWatch;
-            _processorOptions = new ProcessorOptions(runnerContext);
-
-            Processor = processor;
-            RunnerContext = runnerContext;
-
-            var migrationRunnerConventionsAccessor = new AssemblySourceMigrationRunnerConventionsAccessor(
-                serviceProvider: null,
-                new AssemblySource(() => assemblies));
-            Conventions = migrationRunnerConventions ?? migrationRunnerConventionsAccessor.MigrationRunnerConventions;
-
-            var convSet = conventionSet ?? new DefaultConventionSet(runnerContext);
-
-            _migrationScopeManager = migrationScopeHandler ?? new MigrationScopeHandler(Processor);
-            _migrationValidator = new MigrationValidator(_logger, convSet);
-            MigrationLoader = new DefaultMigrationInformationLoader(Conventions, _migrationAssemblies,
-                                                                    runnerContext.Namespace,
-                                                                    runnerContext.NestedNamespaces, runnerContext.Tags);
-            ProfileLoader = new ProfileLoader(runnerContext, this, Conventions);
-            MaintenanceLoader = new MaintenanceLoader(_migrationAssemblies, runnerContext.Tags, Conventions);
-
-            if (runnerContext.NoConnection)
-            {
-                _versionLoader = new Lazy<IVersionLoader>(
-                    () => new ConnectionlessVersionLoader(
-                        this,
-                        _migrationAssemblies,
-                        convSet,
-                        Conventions,
-                        runnerContext,
-                        versionTableMetaData));
-            }
-            else
-            {
-                _versionLoader = new Lazy<IVersionLoader>(
-                    () => new VersionLoader(this, _migrationAssemblies, convSet, Conventions, runnerContext, versionTableMetaData));
-            }
-        }
-
-        /// <summary>
-        /// Initializes a new instance of the <see cref="MigrationRunner"/> class.
-        /// </summary>
         /// <param name="options">The migration runner options</param>
         /// <param name="processorOptions">The migration processor options</param>
         /// <param name="profileLoader">The profile loader</param>
