+++ src/FluentMigrator.Runner/FluentMigratorServiceCollectionExtensions.cs
@@ -1,0 +1,71 @@
+#region License
+// Copyright (c) 2018, FluentMigrator Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#endregion
+
+
+using System;
+using FluentMigrator.Runner.Initialization;
+using FluentMigrator.Runner.Initialization.AssemblyLoader;
+using FluentMigrator.Runner.Initialization.NetFramework;
+using FluentMigrator.Runner.Logging;
+using FluentMigrator.Runner.Processors;
+using JetBrains.Annotations;
+using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Logging;
+
+namespace FluentMigrator.Runner
+{
+    public static class FluentMigratorServiceCollectionExtensions
+    {
+        /// <summary>
+        /// Add all database services to the service collection
+        /// </summary>
+        /// <param name="services">The service collection</param>
+        /// <returns>The service collection</returns>
+        private static IServiceCollection AddAllDatabases(this IServiceCollection services)
+        {
+            return services
+                .ConfigureRunner(
+                    builder => builder
+                        .AddDb2()
+                        .AddDb2ISeries()
+                        .AddDotConnectOracle()
+                        .AddDotConnectOracle12C()
+                        .AddFirebird()
+                        .AddHana()
+                        .AddMySql4()
+                        .AddMySql5()
+                        .AddOracle()
+                        .AddOracle12C()
+                        .AddOracleManaged()
+                        .AddOracle12CManaged()
+                        .AddPostgres()
+                        .AddPostgres92()
+                        .AddRedshift()
+                        .AddSqlAnywhere()
+                        .AddSQLite()
+                        .AddSqlServer()
+                        .AddSqlServer2000()
+                        .AddSqlServer2005()
+                        .AddSqlServer2008()
+                        .AddSqlServer2012()
+                        .AddSqlServer2014()
+                        .AddSqlServer2016()
+                        .AddSqlServerCe());
+        }
+
+    }
+
+}
